split and move observer, observable
